+++ src/Field.h
@@ -16,9 +16,9 @@ private:
 	sf::RectangleShape figure_next[4]; // след. фигура
 	sf::RectangleShape figure_swap[4]; // фигура для смены
 	sf::RectangleShape figure_shadow[4]; // тень фигуры
-	int choice_next = 0; // номер след. фигуры из вектора functions
-	int choice_swap = 0; // номер фигуры для смены
-	int choice_cur = 0; // номер текущей фигуры
+	size_t choice_next = 0; // номер след. фигуры из вектора functions
+	size_t choice_swap = 0; // номер фигуры для смены
+	size_t choice_cur = 0; // номер текущей фигуры
 	float levelSpeed = 0.5f; // скорость падения блоков в зависимости от уровня
 	Blocks blocks; // матрица блоков которые остались на поле
 	sf::RenderWindow& window;
@@ -39,7 +39,7 @@ private:
 	// генераторы случайных чисел для выбора фигуры из вектора functions
 	bool isInit = false;
 	std::mt19937 generator;
-	std::uniform_int_distribution<int> distribution{ 0, 6 }; // всего 7 фигур
+	std::uniform_int_distribution<unsigned int> distribution{ 0, 6 }; // всего 7 фигур
 
 	// функции для создания фигур
 	void createI();
@@ -51,7 +51,7 @@ private:
 	void createZ();
 
 	// перевести реальные координаты в координаты, которые соответствуют полю
-	std::pair<int, int> coordsTofield(const sf::Vector2f& coords);
+	std::pair<uint16_t, uint16_t> coordsTofield(const sf::Vector2f& coords);
 	
 	// есть ли блок на месте с координатами coords: false - есть, true - нету
 	bool checkIntersect(const sf::Vector2f& coords);
@@ -66,7 +66,7 @@ private:
 	void placeSwapFigure(); // разместить swap фигурy в спец. местe за полем
 
 public:
-	Field(int height, int width, sf::RenderWindow& win, sf::Font& font, FieldPosition pos);
+	Field(uint16_t height, uint16_t width, sf::RenderWindow& win, sf::Font& font, FieldPosition pos);
 
 	// отображение поля, блоков, фигуры и счета в окне window 
 	void drawFigure();
@@ -77,9 +77,9 @@ public:
 	void initFigure(); // начальная инициализация фигуры
 
 	// сдвиг фигуры: side=1 - вправо, side=-1 - влево, down=1 - вниз; 0 - сдвига нету
-	void moveFigure(int side = 0, int down = 0); 
+	void moveFigure(int8_t side = 0, int8_t down = 0);
 
-	bool checkSides(int moveSide); // можно ли сделать сдвиг в сторону
+	bool checkSides(int8_t moveSide); // можно ли сделать сдвиг в сторону
 	bool checkDown(); // можно ли двигаться фигуре вниз
 
 	void rotate(bool force = false); // поворот фигуры; force - поворот даже если фигура выйдет за границы
@@ -98,7 +98,7 @@ public:
 
 	void swapFigure(); // сменить фигуру
 
-	void calculateSpeed(int level); // посчитать скорость падения в зависимости от уровня
+	void calculateSpeed(uint8_t level); // посчитать скорость падения в зависимости от уровня
 
 	// тень нужно пересоздавать каждый раз после createNewFigure, rotate, swap, moveSide
 	void createShadow();
@@ -107,6 +107,6 @@ public:
 
 	float getSpeed() { return levelSpeed; }
 
-	void setSeed(unsigned int  seed);
+	void setSeed(uint32_t seed);
 };
